Break up common.js and attempt to prevent "flash" when navigating around
